Revert "83 implement file existence and permission checks"
