+++ includes/lib/utils/string_utils.hpp
@@ -1,7 +1,6 @@
 #ifndef LIB_UTILS_STRING_UTILS_HPP_
 #define LIB_UTILS_STRING_UTILS_HPP_
 
-#include <cctype>
 #include <sstream>
 #include <string>
 #include <vector>
