+++ srcs/ConfigParser.cpp
@@ -12,19 +12,6 @@ ConfigParser::ConfigParser(const std::string& text)
 ConfigParser::~ConfigParser() {
 }
 
-void ConfigParser::LoadFileOrThrowRuntime(const std::string& filename) {
-  struct stat s;
-  if (stat(filename.c_str(), &s) != 0) {
-    throw std::runtime_error("File does not exist: " + filename);
-  }
-  if (S_ISDIR(s.st_mode)) {
-    throw std::runtime_error(filename + " is a directory");
-  }
-  std::ifstream file(filename.c_str());
-  if (!file.is_open()) {
-    throw std::runtime_error("Failed to open file: " + filename);
-  }
-  content = std::string((std::istreambuf_iterator<char>(file)),
-                        std::istreambuf_iterator<char>());
-  file.close();
+void ConfigParser::LoadFile(const std::string& filename) {
+  content = lib::utils::ReadFile(filename);
 }
