+++ srcs/RequestHandler.cpp
@@ -4,7 +4,6 @@
 #include <stdexcept>
 
 #include "CgiExecutor.hpp"
-#include "FileValidator.hpp"
 #include "HttpRequest.hpp"
 #include "ServerConfig.hpp"
 #include "lib/http/Method.hpp"
@@ -56,23 +55,10 @@ std::string RequestHandler::ResolveFilesystemPath() const {
   }
   const std::string req_uri = req_.GetUri();
   std::string path = location_match_.loc->GetRoot() + location_match_.remainder;
-  // Validate the path for security
-  path = FileValidator::ValidateAndNormalizePath(
-      path, location_match_.loc->GetRoot());
-  // TODO: create AppendIndexIfDirectory(path, req_uri, index)
   bool req_uri_ends_with_slash =
       (!req_uri.empty() && req_uri[req_uri.size() - 1] == '/');
   bool is_directory =
       (req_uri_ends_with_slash || lib::utils::IsDirectory(path));
-  /* TODO: replace lib::utils::IsDirectory() with StatOrThrow + S_ISDIR later
-  bool is_directory = false;
-  if (req_uri_ends_with_slash) {
-    is_directory = true;
-  } else {
-    struct stat st = lib::utils::StatOrThrow(path);
-    is_directory = S_ISDIR(st.st_mode);
-  }
-  */
   if (is_directory) {
     if (location_match_.loc->GetIndexFile().empty()) {
       throw std::runtime_error("No index files configured for location");
@@ -88,7 +74,7 @@ void RequestHandler::HandleGet() {
     CgiExecutor cgi(req_, *location_match_.loc, filesystem_path_);
     res_ = cgi.Run();
   } else {
-    std::string body = lib::utils::ReadFileToStringOrThrow(filesystem_path_);
+    std::string body = lib::utils::ReadFile(filesystem_path_);
     res_.AddHeader("Content-Type",
                    lib::http::DetectMimeTypeFromPath(filesystem_path_));
     res_.SetBody(body);
@@ -101,7 +87,7 @@ void RequestHandler::HandlePost() {
     CgiExecutor cgi(req_, *location_match_.loc, filesystem_path_);
     res_ = cgi.Run();
   } else {
-    std::string body = lib::utils::ReadFileToStringOrThrow(filesystem_path_);
+    std::string body = lib::utils::ReadFile(filesystem_path_);
     res_.AddHeader("Content-Type",
                    lib::http::DetectMimeTypeFromPath(filesystem_path_));
     res_.SetBody(body);
@@ -114,7 +100,7 @@ void RequestHandler::HandleDelete() {
     CgiExecutor cgi(req_, *location_match_.loc, filesystem_path_);
     res_ = cgi.Run();
   } else {
-    std::string body = lib::utils::ReadFileToStringOrThrow(filesystem_path_);
+    std::string body = lib::utils::ReadFile(filesystem_path_);
     res_.AddHeader("Content-Type",
                    lib::http::DetectMimeTypeFromPath(filesystem_path_));
     res_.SetBody(body);
