+++ srcs/lib/utils/string_utils.cpp
@@ -1,5 +1,12 @@
 #include "lib/utils/string_utils.hpp"
 
+#include <cctype>
+#include <sstream>
+#include <string>
+#include <vector>
+
+#include "lib/type/Optional.hpp"
+
 namespace lib {
 namespace utils {
