--- tests/file_validater/FileValidator_ContainsDotDotSegments.test.cpp
@@ -1,20 +1,0 @@
-#include <gtest/gtest.h>
-
-#include "FileValidator.hpp"
-
-TEST(FileValidatorTest, ContainsDotDotSegments_RejectsDotDotSegments) {
-  EXPECT_TRUE(FileValidator::ContainsDotDotSegments("/../a"));
-  EXPECT_TRUE(FileValidator::ContainsDotDotSegments("/a/../b"));
-  EXPECT_TRUE(FileValidator::ContainsDotDotSegments("/a/.."));
-  EXPECT_TRUE(FileValidator::ContainsDotDotSegments(".."));
-  EXPECT_TRUE(FileValidator::ContainsDotDotSegments("../a"));
-}
-
-TEST(FileValidatorTest, ContainsDotDotSegments_AcceptsSafePaths) {
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/a..b"));
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/a/..b"));
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/a/b.."));
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/a/b/c"));
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/"));
-  EXPECT_FALSE(FileValidator::ContainsDotDotSegments("/a/b/./c"));
-}
