+++ tests/request_handler/RequestHandler_ResolveFilesystemPath.test.cpp
@@ -21,10 +21,8 @@ TEST_F(RequestHandlerTest, ResolveFilesystemPath_RootDirectory) {
     HttpRequest request;
     request.SetUri("/");
     RequestHandler handler(config, request);
-
-    std::string out;
-    ASSERT_NO_THROW(out = handler.ResolveFilesystemPath());
-    EXPECT_EQ(out, "/var/www/html/index.html");
+    
+    EXPECT_EQ(handler.ResolveFilesystemPath(), "/var/www/html/index.html");
 }
 
 TEST_F(RequestHandlerTest, ResolveFilesystemPath_SubDirectory) {
